+expr bracket support
